products money

+++ utils/validation/productsValidator.ts
@@ -3,6 +3,7 @@ import { check } from "express-validator";
 import validatorMiddleware from "../../middlewares/validatorMiddleware";
 import categoriesModel from "../../models/categoriesModel";
 import productsModel from "../../models/productsModel";
+import shopsModel from "../../models/shopsModel";
 import { CategoryModel, ProductModel } from "../../interfaces";
 
 export const createProductValidator = [
@@ -26,8 +27,8 @@ export const createProductValidator = [
     check('category')
         .notEmpty().withMessage('category is required')
         .isMongoId().withMessage('Invalid category id')
-        .custom(async (categoryId: string): Promise<boolean> => {
-            const category: CategoryModel | null = await categoriesModel.findById(categoryId);
+        .custom(async (categoryId: string, { req }): Promise<boolean> => {
+            const category: CategoryModel | null = await categoriesModel.findOne({ _id: categoryId, shop: req.user.shop });
             if (!category) { return Promise.reject(new Error('Category does not exist')); };
             return true;
         }),
@@ -43,14 +44,22 @@ export const updateProductValidator = [
     check('id').isMongoId().withMessage("invalid product id"),
     check('name').optional().isLength({ min: 2, max: 50 }).withMessage("name length must be between 2 and 50"),
     check('description').optional().isLength({ min: 2, max: 300 }).withMessage("description length must be between 2 and 300"),
-    check('quantity').optional().isNumeric().withMessage('Quantity Must be a number').toInt(),
     check('productPrice').optional().isNumeric().withMessage('product price Must be a number').toFloat(),
     check('sellingPrice').optional().isNumeric().withMessage('selling price Must be a number').toFloat(),
     check('sold').optional().isNumeric().withMessage('sold Must be a number').toInt(),
     check('category').optional().isMongoId().withMessage('Invalid category id')
-        .custom(async (categoryId: string): Promise<boolean> => {
-            const category: CategoryModel | null = await categoriesModel.findById(categoryId);
+        .custom(async (categoryId: string, { req }): Promise<boolean> => {
+            const category: CategoryModel | null = await categoriesModel.findOne({ _id: categoryId, shop: req.user.shop });
             if (!category) { return Promise.reject(new Error('Category does not exist')); };
+            return true;
+        }),
+    check('quantity').optional().isNumeric().withMessage('Quantity Must be a number').toInt()
+        .custom(async (quantity: number, { req }) => {
+            if (quantity < 0) { return Promise.reject(new Error('Quantity must be greater than 0')); };
+            const product: ProductModel | null = await productsModel.findById(req.params?.id)
+            if (!product) { return Promise.reject(new Error('Product Not found')); };
+            const compareQuantity: number = quantity - product.quantity;
+            await shopsModel.findByIdAndUpdate(product.shop, { $inc: { productsMoney: compareQuantity * product.productPrice } }, { new: true });
             return true;
         }),
     validatorMiddleware
@@ -73,12 +82,14 @@ export const deleteProductValidator = [
     check('id').isMongoId().withMessage("invalid product id")
         .custom(async (id: string): Promise<boolean> => {
             const product: ProductModel | null = await productsModel.findById(id);
+            if (!product) { return Promise.reject(new Error('Product Not found')); };
             const productImages: string[] = [];
             product?.images.forEach((imageUrl: string): void => {
                 const image: string = imageUrl.split(`${process.env.Base_URL}/products/`)[1];
                 productImages.push(image);
             });
             deleteUploadedImages(productImages);
+            await shopsModel.findByIdAndUpdate(product.shop, { $inc: { productsMoney: -product.quantity * product.productPrice } }, { new: true });
             return true;
         }),
     validatorMiddleware
